+++ packages/examples/packages/file-upload/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './UploadedFiles';
+export * from './UploadForm';