+++ packages/snaps-jest/src/internals/simulation/interface.ts
@@ -323,6 +323,7 @@ export async function clickElement(
         type: UserInputEventType.FormSubmitEvent,
         name: result.form,
         value: state[result.form] as Record<string, string | null>,
+        files: {},
       },
       context,
     );