+++ packages/snaps-sdk/src/jsx/component.ts
@@ -40,6 +40,17 @@ export type SnapElement<
 };
 
 /**
+ * A type that can be a single value or an infinitely nestable array of values.
+ *
+ * @template Type - The type that can be an array.
+ * @example
+ * type NestableString = Nestable<string>;
+ * const nestableString: NestableString = 'hello';
+ * const nestableStringArray: NestableString = ['hello', 'world', ['foo', ['bar']]];
+ */
+export type Nestable<Type> = Type | Nestable<Type>[];
+
+/**
  * A type that can be a single value or an array of values.
  *
  * @template Type - The type that can be an array.
@@ -48,7 +59,7 @@ export type SnapElement<
  * const maybeArrayString: MaybeArrayString = 'hello';
  * const maybeArrayStringArray: MaybeArrayString = ['hello', 'world'];
  */
-export type MaybeArray<Type> = Type | Type[];
+export type MaybeArray<Type> = Nestable<Type>;
 
 /**
  * A JSX node, which can be an element, a string, null, or an array of nodes.