+++ packages/snaps-sdk/src/jsx/components/form/FileInput.test.tsx
@@ -1,0 +1,29 @@
+import { FileInput } from './FileInput';
+
+describe('FileInput', () => {
+  it('renders a file input', () => {
+    const result = <FileInput name="foo" accept={['image/png']} />;
+
+    expect(result).toStrictEqual({
+      type: 'FileInput',
+      props: {
+        name: 'foo',
+        accept: ['image/png'],
+      },
+      key: null,
+    });
+  });
+
+  it('renders a compact file input', () => {
+    const result = <FileInput name="foo" compact />;
+
+    expect(result).toStrictEqual({
+      type: 'FileInput',
+      props: {
+        name: 'foo',
+        compact: true,
+      },
+      key: null,
+    });
+  });
+});