+++ packages/snaps-sdk/src/jsx/components/form/index.ts
@@ -1,6 +1,7 @@
 import type { ButtonElement } from './Button';
 import type { DropdownElement } from './Dropdown';
 import type { FieldElement } from './Field';
+import type { FileInputElement } from './FileInput';
 import type { FormElement } from './Form';
 import type { InputElement } from './Input';
 import type { OptionElement } from './Option';
@@ -9,6 +10,7 @@ export * from './Button';
 export * from './Dropdown';
 export * from './Option';
 export * from './Field';
+export * from './FileInput';
 export * from './Form';
 export * from './Input';
 
@@ -16,6 +18,7 @@ export type StandardFormElement =
   | ButtonElement
   | FormElement
   | FieldElement
+  | FileInputElement
   | InputElement
   | DropdownElement
   | OptionElement;