+++ packages/test-snaps/src/features/snaps/file-upload/constants.ts
@@ -1,0 +1,5 @@
+import packageJson from '@metamask/file-upload-example-snap/package.json';
+
+export const FILE_UPLOAD_SNAP_ID = 'npm:@metamask/file-upload-example-snap';
+export const FILE_UPLOAD_SNAP_PORT = 8030;
+export const FILE_UPLOAD_VERSION = packageJson.version;